requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
streamlit>=1.36.0
pandas>=2.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
python-dotenv>=1.0.0
requests>=2.32.0
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.7
langchain-community>=0.2.0
faiss-cpu>=1.8.0
tiktoken>=0.7.0
